SECURITY: fix race condition in path check
chroot only continues if all direcotries in the given path are owned by
the superuser and none are writable by the group or others. But these
checks are only executed on the resolved canonical path, not the path
originally passed by the caller that is used to chroot (and which may
contain symlinks owned by a non-privileged attacker). This opens up a
race condition between the check of the resolved path and the call to
chroot and allows an attacher to quickly change symlinks in between.

Solved by using the resolved canonical path to chroot.