SECURITY: convert group names to numeric ids before chroot
Convert user and group names to numeric ids before entering the
chroot. This ensures /etc/passwd and /etc/group of the original
file-system are used, and not from within the chroot. Also make
sure the eventual user and group id are not 0, whatever the name may
be. Furthermore, do this check before chrooting, not only after.

This is considered to be a security issue because although chroot only
continues if all components in the canonical path are owned by the
superuser and none are writable by the group or others, these checks are
not done on the non-canonical path that is actually used to chroot and
thus there is a race condition that will be fixed in a subsequent commit.

Special thanks to hubnut for pointing this out.